requirements.txt
ollama
load_dotenv
discord.py[voice]
audioop-lts
sqlalchemy